move rest of decision logic into decision module
