+++ reconcile/test/test_change_owners.py
@@ -6,7 +6,6 @@ from reconcile.change_owners.diff import (
     deepdiff_path_to_jsonpath,
 )
 from reconcile.change_owners.change_owners import (
-    apply_decisions_to_changes,
     manage_conditional_label,
 )
 from reconcile.change_owners.self_service_roles import (
@@ -14,6 +13,7 @@ from reconcile.change_owners.self_service_roles import (
 )
 from reconcile.change_owners.decision import (
     get_approver_decisions_from_mr_comments,
+    apply_decisions_to_changes,
     DecisionCommand,
     Decision,
 )
